refactor: move order quote function to its own file
