+++ tauri-app/src-tauri/src/commands/mod.rs
@@ -5,6 +5,7 @@ pub mod config;
 pub mod dotrain;
 pub mod dotrain_add_order_lsp;
 pub mod order;
+pub mod order_quote;
 pub mod order_take;
 pub mod vault;
 pub mod wallet;
