+++ tauri-app/src-tauri/src/main.rs
@@ -13,9 +13,10 @@ use commands::config::{convert_configstring_to_config, merge_configstrings, pars
 use commands::dotrain::parse_dotrain;
 use commands::dotrain_add_order_lsp::{call_lsp_completion, call_lsp_hover, call_lsp_problems};
 use commands::order::{
-    batch_order_quotes, compose_from_scenario, order_add, order_add_calldata, order_detail,
-    order_remove, order_remove_calldata, orders_list, orders_list_write_csv,
+    compose_from_scenario, order_add, order_add_calldata, order_detail, order_remove,
+    order_remove_calldata, orders_list, orders_list_write_csv,
 };
+use commands::order_quote::batch_order_quotes;
 use commands::order_take::{order_takes_list, order_takes_list_write_csv};
 use commands::vault::{
     vault_balance_changes_list, vault_balance_changes_list_write_csv, vault_deposit,
